Add files via upload
In this project our goal of the project is to understand what leads a customer to churn. We clearly want our valued customers to stay with us as long as possible, and we want their experience to be wonderful. Therefore, we want to allocate our resources to the analysis of what may lead customers to no longer come to us for business.